create explainTab, tabStatusIndicator, queryResultTabs components
